Internal:  - Enabled code map for VS2013 files (if you don't use it, it does nothing so no harm in turning it on)  - Ignored some new project files (ipch folder and *.sdf files)  - Fixed warning in BR_Loudness.cpp

+++ Breeder/BR_Loudness.cpp
@@ -524,7 +524,7 @@ void BR_LoudnessObject::GotoMomentaryMax (bool timeSelection)
 		return;
 
 	double newPosition = (m_track) ? (0) : (GetMediaItemInfo_Value(GetMediaItemTake_Item(m_take), "D_POSITION"));
-	for (int i = 0; i < m_momentaryValues.size(); ++i)
+	for (size_t i = 0; i < m_momentaryValues.size(); ++i)
 	{
 		if (m_momentaryValues[i] == m_momentaryMax)
 			break;
@@ -553,7 +553,7 @@ void BR_LoudnessObject::GotoShortTermMax (bool timeSelection)
 		return;
 
 	double newPosition = (m_track) ? (0) : (GetMediaItemInfo_Value(GetMediaItemTake_Item(m_take), "D_POSITION"));
-	for (int i = 0; i < m_shortTermValues.size(); ++i)
+	for (size_t i = 0; i < m_shortTermValues.size(); ++i)
 	{
 		if (m_shortTermValues[i] == m_shortTermMax)
 			break;
